+++ CHANGELOG.md
@@ -2,10 +2,25 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [40.2.0] 2026-03-12
+## [41.1.0] 2026-03-24
 ### Added
 - Support for rendering-markdown extension. Note that rendering-xhtml
   takes precedence over rendering-markdown if both are present.
+
+## [41.0.0] 2026-03-13
+### Changed
+- If answerExpression updates the answer list, we do not reset item.value
+  to null. Instead, any values that are still valid with the new answer
+  list will be kept, and any values that are no longer valid will be removed.
+  This is a breaking change.
+- For a checkbox layout question with sub items, if the answer list is
+  updated by answerExpression, the sub items associated with the now-invalid
+  value will also be removed.
+### Fixed
+- For a checkbox layout question with sub items, the sub items will
+  be rendered for each selected checkbox option.
+- When exported to QuestionnaireResponse, each sub item will go under its
+  associated answer, in the format of item.answer.item.
 
 ## [40.1.3] 2026-03-12
 ### Added